requirements.txt
opencv-python
Pillow
scikit-learn
tqdm
torch
torchvision
wandb
fastapi
uvicorn
python-multipart